Initial commit.
This is the basic infrastructure needed to build the Linux daemon
needed for the project codenamed Wired-Leaf. It includes a skeletal
.proto, which gives the flavor of what we expect to be sending to and
receiving from open-ephys, and the SCons build infrastructure needed
to get protobuf-c sources generated and linked against by the daemon
sources.

For now, the daemon is just a program that writes a serialized
protobuf to stdout.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>