Have daemon attempt to read some data node config.
Assume there's a client at localhost:8880, which can reply to basic
system configuration requests. Ask it for the information, log it, and
quit.

This example makes it pretty clear that there needs to be more
raw_packets machinery for wrapping up request/response pairs in a
convenient fashion, but leave it as is for now, as it's enough for a
test.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>