src/main.c: Add primitive resistance to UDP errors.
Add a timeout (100 msec) instead of blindly calling recv for each
datanode board sample packet. Read as many packets as are available,
and ignore results with unexpected indexes. Re-send requests on
timeout. Barring protocol errors or error flags getting set (which the
dummy node currently doesn't commit), that should suffice to ensure we
get all the packets.

We can now move on to basic file reading and retrieval using protobuf,
but that will require multithreading and refactoring if it's not going
to be a mess.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>