dummy-datanode: Randomly drop/duplicate board samples.
In the spirit of Chaos Monkey, have dummy-datanode behave badly
according to a configurable set of probabilities, with an API in
chaos.h to control the overall behavior.  The current rates at which
mischief occurs (defined in chaos.c) are intentionally very
high. Decrease the number of fake samples to make dealing with this
tolerable.

You can run dummy-datanode against the daemon and confirm for yourself
that the samples eventually do get there. (Of course, this doesn't
confirm that the samples arrived unharmed, but that's a test for
another patch series.)

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>