daemon: write received packets to persistent storage.
Rip out the data storage code from main.c, and put it behind an
abstraction for channel storage defined in ch_storage.h.  It's based
on a struct ch_storage, which works similarly to Linux VFS. We use
this abstraction in src/main.c to store packets received from the data
node.

The details of the storage backend are hidden behind ch_storage
alloc/dealloc routines in raw_ch_storage.h and hdf5_ch_storage.h. The
HDF5 backend doesn't work yet; several of its routines are just stubs
that return errors.

I'm not over-excited about the extra indirection, but it seems like
the only reasonable choice if we want to both hide the details of
packet persistence from the main daemon thread and allow for
convenient backend substitution.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>