lib: raw_packets: Add raw_packet_recvmmsg().
We need a clean way to request and then read multiple board sample
packets at once. We could send the request for multiple packets, then
read them out one datagram at a time. That's likely to be inefficient,
though; it's better to read multiple datagrams in a single system call
with recvmmsg().

To make this easier, raw_packet_recvmmsg() is a convenience wrapper
around recvmmsg() which createsa and initializes the struct mmsghdr
for you. It doesn't let you see per-msghdr flags that get returned,
though. If you want that, you must do it yourself for now.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>