raw_packets: Add note about undefined behavior.
It's technically not OK to embed a struct with a flexible array member
inside a union contained in another struct like this, but I'm not too
concerned about this messing us up. The unit tests should catch any
future compiler releases that don't implement this in the obvious way,
and it makes the rest of the packet-processing code a lot easier to
write.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>