lib: Replace raw_packets to reflect new protocol.
This reflects the current docs/datanode-net-protocol-notes.txt. The
new API is similar to the old one.

We need more datanode spec before this can be usable to implement a
real-ish dummy session, though. In particular, the register interfaces
for setting up, conducting, and terminating a recording session aren't
written yet.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>