Support reg. I/O for well-behaved control sessions
NOTE: We temporarily now require that the data node connect to the
      daemon. A patch to dummy-datanode which makes this happen will
      follow next. Further patches that tell the daemon how to connect
      to the data node will be needed to restore the old behavior.

You can now use exchange-protobuf to send a protocol message
containing a RegisterIO to the daemon, which will turn it into a
raw_pkt_cmd, forward it to the dummy datanode, wait for the response,
turn the response back into protobuf, and send it back.

main.c: use the control session API in control.h to listen for and
handle control socket traffic. Also install SIGTERM and SIGINT
handlers.

New files and their purposes:

- control.h: Top-level control session API used by main.c. Everything
  else is internal.

- control.c: libevent plumbing and control socket worker thread

- control-client.h, control-client.c: Client-side control work

- control-dnode.h, control-dnode.c: Data node-side control work

- control-private.h: Internal control declarations, shared between
  core control.c and client/data-node pieces.

  This defines the core struct control_session. It also defines struct
  control_ops, which declares a set of function pointers that the
  client and data node code provide to control the behavior of the
  core code in control.c.

Delete data_node.h and data_node.c; the contents of those files is now
a failed experiment.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>