Rework control processing for multiple requests per command.
Add struct control_txn to control-private.h. This encodes a single
request/response pair that must be performed as a result of a
ControlCommand received from the client control socket. Add fields to
struct control_session to allow client code to store an array of
control transactions to perform after receiving a command.

Have control-client.c and control-dnode.c respect this new way of
communicating with each other. RegisterIO commands work in this way.

Along the way, we run into a variety of new ways to mess up, so add
more error handling and error responses to control.proto so we can
indicate to the client when something went wrong.

We'll use this new infrastructure to support data socket setup and
teardown.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>