Add protobuf support for BoardSample/BoardSubsample.
control.proto: add ControlCmdStream, allow it to be embedded in a
ControlCommand to configure the live stream. Add "SUCCESS" type to
ControlResponse, which the daemon can return to indicate successful
stream configuration.

data.proto: new file; contains definitions for Sample messages, each
of which contains either a BoardSubsample or BoardSample message.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>