Revert "Pass ch_storage to sample_session."
This reverts commit 6ba2b697f091481ec835e1c58c5993593b490539, which
was poorly thought through. The channel storage structure should
instead be passed to new sample.h API each time a set of board samples
needs to be retrieved. That will let client control socket callbacks
set up and reconfigure a struct ch_storage each time the user starts
and stops an experiment without worrying that the sample callbacks are
messing with it.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>