sample: add sample_expect_bsamps(), sample_reject_bsamps().
sample_expect_bsamps() allows the control protocol handlers to notify
the sample socket handler that board sample packets are on the way. It
uses a new struct sample_bsamp_cfg to describe the incoming packets,
and to provide a struct ch_storage for storing them.

On the back-end, the sample storage handler receives into a
double-buffered array of board sample packets, and uses a worker
thread to do the actual ch_storage_*() calls (since storage backends
only do synchronous I/O).

sample_reject_bsamps() is a quick-exit feature which background
threads can use to abort an ongoing transfer prematurely. (It blocks
if the sample worker thread is currently storing samples, so it's not
smart to call it from event loop callbacks.)

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>