control.proto: add protocol messages for sample storage
These allow clients to specify capture and storage to disk of either
live or previously-recorded board samples, to either HDF5 or as raw
packets.

Obviously, that's a pretty nasty security hole with an untrusted
client. We haven't worried about that for any of the other innumerable
places this daemon can't be exposed to the public Internet, so there's
no reason to start now.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>