Add test/test-libsng.
This tests that the shared library interface to the daemon we're
providing to SNG works as advertised. Assuming the daemon doesn't drop
any packets and that there are no network errors, the tests pass.

You can't run it unless you put the generated libsng into somewhere
your dynamic linker will find it. E.g. on Linux,

$ sudo cp build/libsng/libsng.so /usr/lib

That's an annoyance, but I don't want to add the SCons install targets
to automate the process, since I'm happier if building and installing
the daemon doesn't require root, and I'm apparently too much of a n00b
to figure out how to get LD_PRELOAD to work. I suspect malformed
shared library dependencies in the generated libsng. We can sort this
out later when it comes time to deploy this.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>