control-client.c: tweak comments about UDP race
The problem described is the same as the one solved by TCP's TIME_WAIT
state. Write that down in the comments. We can add that functionality
in if we have to, but I see no reason to reimplement TCP on top of UDP
right now, so we'll just let the issue stand.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>