control-client.c: band-aid for close or stop during sample storage.
Halt any ongoing transfers on client_close() or client_stop(). Doing
it during client_close() blocks the event loop thread while the sample
worker thread finishes its current write. Fixing that would be a bit
invasive, though, and it's not an expected case, so just mark it as a
FIXME.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>