Support disk->disk sample storage.
Add bs_restarted field to struct client_priv; when bs_expecting is
nonzero, bs_restarted is a predicate for whether the currently
expected board samples are due to a restarted storage after a dropped
packet. We handle restarts by tweaking the original ControlCmdStore;
since the altered command didn't originate from the user, we need to
check if what we're handling is due to a restart to e.g. avoid sending
error results prematurely.

Also add bs_nwritten_cache, when we need to restart, this keeps track
of the total number of samples written to disk over all restarted
storage operations.

Factor out the code in client_sample_store_callback() which sent the
results of a ControlCmdStore into its own routine,
client_send_store_res(). We need this in a couple of places. In its
place, add the necessary logic to update the struct sample_bsamp_cfg
and signal to the worker thread that it needs to handle the updated
command.

Add the necessary protobuf-to-register-transactions glue. The control
transactions are based on the sata_read() command in the sng-mcu-tcp
repo as of 27a20d159ba930ae195c2c6b59b9ed04ec714033, with a little
extra advice from Bryan.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>