lib: add safe_pthread.h
This contains wrappers for various bits of the pthread API that turn
all errors fatal. We've duplicated that functionality in a few places,
so we'll move those to using this API next.

Including safe_pthread.h with SAFE_PTHREAD_DEBUG_LOG causes the
function which called any of its routines to be logged, along with the
stringification of the arguments passed.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>