Add support for reconnecting to the data node.
Add a background thread that reconnects to the data node when the
connection gets dropped. It notifies the main thread when the
connection succeeds by activating an event. This works identically to
the client evconnlistener callback.

This requires changing the semantics of the client and data node open
methods. These are now called with the main control session lock held.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>