Move various utility programs to util/
There's lots of bits of utility functionality that should be
consolidated.

Move the following subdirectories of test/ to util/:

     dummy-datanode
     proto2bytes
     sampstreamer

Update SConstruct so they still get built and their programs placed
under build/, as usual.

Move the top-level exchange-protobuf script to util/. Fix the path in
a reference to it in control.proto.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>