Add test/run-tests.py.
This runs all of the tests. Add an Install to SConstruct which copies
it to the build directory.

run-tests.py MUST be called as the installed version in the build
directory. (Tests rely on compiled utility programs, and some tests
are written in C, so it doesn't make sense to run it from test/.)

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>