sockutil: add sockutil_ntop()
I really don't understand why inet_ntop() is so stupid.

It forces you to know the socket structure in advance, so you know the
offset of the address inside, which ruins the point of having
something that works for multiple address families. Every user of
inet_ntop() needs to know about any address family that might exist,
so if we ever get IPv(n+1), we'll have to revisit all of those call
sites.

sockutil_ntop() fixes this for you, by putting all the knowledge about
address structures in one place, especially when combined with
sockutil_addrstrlen().

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>