control-client.c: change confusing cpriv field names
"c_cmdlen" is confusing, since that variable really measures the
number of bytes on the wire of c_cmd's protocol buffer; it's not the
length of the c_cmd protocol message itself (which makes no sense).

Accordingly, rename c_cmdlen to c_pbuflen, since it's the number of
bytes to read to fill c_pbuf. Rename its associated evbuffer,
c_cmdlen_buf, to c_pbuflen_buf.

This entire thing is pretty stupid; we should be using bufferevent
watermarks anyway. This was some of the early libevent-aware code in
the project, though, so it's kind of naively implemented. It works and
isn't performance-critical, though, so just leave it as-is for now.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>