Add timeout for data node transactions.
When a data node transaction takes too long (here, 1.75 seconds),
assume that everything is hosed, and close the data node
connection. If matters improve, later commands can succeed after the
connection has been re-established.

When this occurs, use cs_partner_closed() in control-client.c to send
the client a response indicating the failure (a ControlResponse with
error type, and .err field with code DNODE_DIED). This resolves a
FIXME.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>