control-client.c: don't restart failed storage forever
If the data node is really hosed (e.g. if it missed writing a packet
to disk), we may try to perform a storage request where at a certain
start index, we'll never succeed in writing any more packets.

Rather that have this lead to an infinite loop, detect if this happens
too often (here 20 times), and if so, don't bother with further
restarts.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>