control-client.c: fix race condition on early dropped packet.
We generally shit the bed if sample storage fails while we're waiting
for a result. That was wrong to start with, and is causing assertion
failures in live testing.

Fix that by checking if we're expecting a response in the sample
callback handler. If we are, wait for the response to come back before
moving on and replacing the current transactions with those
corresponding to the restarted storage. The "response received"
notification comes to control-client.c via the background thread, so
we need to do some extra bookkeeping and libevent plumbing to handle
this properly.

If the response times out, the timeout handler will cause the data
node connection to reopen. We'll be informed about that in
client_partner_closed, so use that to halt the restart and let the
client know what happened.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>