(BREAKING CHANGE) streaming behavior updates
Update streaming behavior to not reset the DAQ module by default. To
get the old behavior, you must now set force_daq_reset=True in your
ControlCmdStream.

This breaks the test for ControlCmdStore, which relies on the old
behavior. Have the test start/stop acquisition around the store
command, as we now recommend in this situation.

Updated tests pass.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>