test_helpers.py: usability fixups for DaemonTest.setUp()
Use the IOError argument to describe what went wrong when we need to
error out at setUp() time.

Reduce number of PING_DNODE attempts to 8. This is more than enough
when things are working properly; the current value of 100 makes the
tests take too long to fail if something's wrong.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>