Add util/expand_eth_buffers.sh.
Copied from sng-mcu-tcp.

This is a shell script which expands ethernet buffers, and
reconfigures the eth0 endpoint MTU to 4096 bytes.

At present time, you must run this before using the daemon. (The
daemon would need to be setuid root if it could handle this itself,
which it's not ready for.)

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>