Modify the way data is saved to HDF5 for quick read access.
Previously, board sample packets were saved to HDF5 in a single, 1D
dataset of compound datatype. This made array slicing impossible, which
forced data analysis routines to loop through the dataset and extract
e.g. the relevant channel data, which is slow. Now, the various parts
of the old compound datatype are saved in separate, contiguous datasets:

* ph_flags
* sample_index
* chip_live
* channel_data
* aux_data

This speeds up the reading of datasets dramatically, and also enables
array slicing.

The attributes which were previously attached to the "wired-dataset"
object are now attached to the root group of the HDF5 file.

Signed-off-by: Chris Chronopoulos <chrono@leaflabs.com>