Avoid use-after-free by reordering cleanup of bsamp_bufs
In sample_finished_with_bsamps(), we were freeing smpl->bsamp_bufs[0]
and smpl->bsamp_bufs[1] to the heap and then accessing fields in these
structs.

The present commit moves freeing of smpl->bsamp_bufs[0] and
smpl->bsamp_bufs[0] to a later position in the function, after access
to these structs is complete.