hdf5: support multiple distinct datasets
Summary:
Currently, a single compound datatype is written as a 1xN dimensional
array of board sample.

In an effort to support more efficient processing, we'd like to split
the board sample data into multiple distinct datasets.

This changeset lays in some infrastructure for creating and writing to
various datasets, based on the contents of the dset_info table.

As samples come in, we copy board sample data through the use of a
scatter function to an intermediate buffer, then create an appropriate
hyperslab to write the board sample data. This buffer is sized to
support a maximum of CHUNK_DIM samples.

As an example, a single dataset, the 'sample_index' dataset, is created.

The existing 'wired-dataset' is left in place. It will be removed once
all the previous data has been migrated to the new data format.

Depends on D77

Test Plan: manual inspection of hdf5 snapshot generated via acquire.py

Reviewers: #willow, tasman

Reviewed By: tasman

Subscribers: tasman, chrono

Differential Revision: https://reviews.leaflabs.com/D78