subsampleemu: add a subsamples stream emulator
Summary:
Add subsampleemu, a tool that writes an emulated subsamples stream to
stdout, as would be output by proto2bytes.

In the initial version of this tool, the emulated stream contains a
12 bit-excursion 1 Khz sine wave in channel 0, with channels 1
through 31 containing the value 0.  The samples are represented as
16-bit unsigned integers, to match the sample format of the Willow
datanode.

Test Plan:
Build willow-daemon, including subsampleemu, via scons.  Start
run willow-stream-analysis/Sonifier/main with the path to subsampleemu
(builds under willow-daemon/build) as an argument.  Sonify channel 0
and confirm that a 1 Khz sine wave is audible.

Reviewers: #willow, perry

Reviewed By: #willow, perry

Differential Revision: https://reviews.leaflabs.com/D285