Retry interrupted socket send and recv calls
Summary:
Add wrapper functions that retry socket send and recv calls that
fail with errno == EINTR.

On Linux, system calls can and do fail with errno == EINTR,
e.g. if interrupted due to arrival of a signal.  Python 3.5
retries interrupted system calls automatically, but in
Python 2.7, the application is responsible for detecting
and retrying such calls, if desired.

Prior to this commit, an interrupted recv call would cause
do_control_commands to give up waiting for a response from
leafysd and issue the next command prematurely, breaking the
"one command at a time" rule mandated by the Willow stack.
leafysd would detect this rule violation, and close WillowGUI's
control socket, severing the link between GUI and daemon.

The bug manifested by having a streaming plugin issue a command
or commands, and then immediately closing its window.  As part
of the Qt subprocess termination, WillowGUI's process received
a signal, and this broke execution of one of the pending commands.
A remaining issue, not covered by the present commit, is that
WillowGUI may still continue processing commands received from a
streaming plugin after that plugin exits, but at least processing
should now complete eventually, with WillowGUI remaining in contact
with leafysd.

Test Plan:
Start WillowGUI and the Chipstream streaming plugin.  Queue a bunch of
commands for the GUI by quickly advancing the chip number via multiple
clicks on the forward arrow to the right of "Plot chip specified" in
in the Chipstream window, then immediately close the Chipstream window.
Observe that the no errors log to the console, and that the WillowGUI
window eventually ungreys, with health status at the top colored green.

Reviewers: #willow, cblam

Reviewed By: #willow, cblam

Differential Revision: https://reviews.leaflabs.com/D307