sockutil: add sockutil_set_tcp_nodelay
Summary:
Add a utility function to set the TCP_NODELAY option on a
SOCK_STREAM socket.  This option configure the socket to
transmit a TCP segment as soon as data is available,
rather than waiting to see if more data arrives.

Test Plan:
For this commit, confirm error-free build of willow-daemon
using 'scons'.  A follow-on commit will use and hence test
the new utility function.

Reviewers: #willow, perry, sinback

Reviewed By: #willow, sinback

Subscribers: sinback

Differential Revision: https://reviews.leaflabs.com/D380