control: configure TCP_NODELAY on sockets
Summary:
Call sockutil_set_tcp_nodelay on the datanode and client
control sockets.  This configures the sockets so that a TCP
segment will be transmitted as soon as data is available,
rather than potentially waiting for the availability of
additional data.

Depends on D380.

Test Plan:
Start WillowGUI and uncheck Watchdog at the top of the window.  Start
capturing  frames on the Ethernet interface connected to the datanode,
e.g. using Wireshark.  Start Chipstream from WillowGUI.  Observe the
captured frames and their timestamps, which represent one chip change
operation, comprising 32 RegisterIO operations.  Compare the timing
with and without the present commit applied, and confirm that
elapsed time to complete the chip change with this commit applied
is less than or equal to the time without the commit.  Full benefit
is observed when applying D380 and the present commit along with
D379.

Reviewers: #willow, perry, sinback

Reviewed By: #willow, sinback

Subscribers: sinback

Differential Revision: https://reviews.leaflabs.com/D381